examples/langchain/instrumentation-comparison/braintrust-eval/pyproject.toml
[project]
name = "braintrust-eval-example"
version = "0.1.0"
description = "Braintrust evaluation with traces sent to AgentOps"
requires-python = ">=3.10"
dependencies = [
    "braintrust[otel]>=0.3.1",
    "autoevals>=0.0.90",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "langchain-aws>=0.2.0",
    "langchain-core>=0.3.0",
]
